version 0.9.3
This is the KagglEliminator.

The staging area is now put to limited, but more efficient use.
The initialization for preset digits now dove tails into the staged update function.  That's it.
Performance is equal to prior results.
Performance on the kaggle data set is excellent, as expected, as the staged update
does it all from start to finish.