version 9.4
add W-Wings to OPT_BIVALS option.
W-Wings perform/yield on the level of Y-Wings.
added a framework for triads/bands and conversion therein.
removed hidden single search for boxes - apparently unnecessary since
triad processing turns them into hidden column singles.
hidden single search for rows can be optionally removed in the same fashion,
except that it is unclear which is better and faster.
This is a noticeable speed increase.